Fixing test packages
